(typegraphql) move list input types to main type file, include user_id in deleteListsById and improve typing and variable names
