+++ backend/graphql/types/List.ts
@@ -1,5 +1,25 @@
-import { Field, ObjectType } from "type-graphql";
-import { Term } from "./Term";
+import { Field, InputType, ObjectType } from "type-graphql";
+import { Term, TermWithoutIds } from "./Term";
+
+@ObjectType()
+@InputType("NewListWithTermsInput")
+/**
+ * Represents a list, with terms, but without userId. To be used mainly
+ * with`createList()`
+ */
+export class NewListWithTerms {
+   @Field()
+   name: string;
+
+   @Field()
+   from_language: string;
+
+   @Field(() => String)
+   to_language: string;
+
+   @Field(() => [TermWithoutIds])
+   terms: TermWithoutIds[];
+}
 
 @ObjectType()
 export class List {
@@ -26,19 +46,24 @@ export class List {
 }
 
 @ObjectType()
-export class MaybeList {
-   @Field(() => List, { nullable: true })
-   list?: List;
-
-   @Field({ nullable: true })
-   error?: string;
-}
-
-@ObjectType()
 export class ListAndTerms {
    @Field(() => List)
    list: List;
 
    @Field(() => [Term], { nullable: "items" })
    terms: Term[];
+}
+
+@ObjectType()
+@InputType("ListUpdateActionInput")
+export class ListUpdateAction {
+   @Field(() => String)
+   type: "name";
+}
+
+@ObjectType()
+@InputType("ListUpdatePayloadInput")
+export class ListUpdatePayload {
+   @Field()
+   name: string;
 }
